requirements.txt
pytorch-lightning==2.1.2
pytorch-lightning==2.1.2
gradio==3.41.2
huggingface-hub
einops
omegaconf
torchmetrics
webdataset
accelerate>=0.26.0
tensorboard
PyMCubes
trimesh
rembg
transformers>=4.36.0
diffusers>=0.24.0
peft>=0.7.0
bitsandbytes
imageio[ffmpeg]
xatlas
plyfile
git+https://github.com/NVlabs/nvdiffrast/